% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variables.R
\name{compare_variables}
\alias{compare_variables}
\title{compare_variables}
\usage{
compare_variables(script1, script2, variables_to_compare)
}
\arguments{
\item{script1}{}

\item{script2}{}
}
\value{
a list
}
\description{
compare_variables
}
