% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variables.R
\name{compare_variables}
\alias{compare_variables}
\title{compare_variables}
\usage{
compare_variables(
  reference,
  student,
  variables_to_compare,
  variable_weights,
  ...,
  return_expected = FALSE
)
}
\arguments{
\item{reference}{A path to a script (reference)}

\item{student}{A path to another script (to test)}

\item{variables_to_compare}{Optional, selected variables to compare}

\item{variable_weights}{Weights to assess the variables. 1 point by default for each variable.}

\item{return_expected}{If a summary table with expected and actual scores should be returned. (See Value)}
}
\value{
If \code{return_expected = FALSE} A list of numerics corresponding to points assigned to each variable (1 each if no weights given).
If \code{return_expected = TRUE} A data frame with colnames "student", "expected" and "score";
}
\description{
compare_variables
}
\examples{
  compare_variables("reference.R", "student_assignment.R")
  compare_variables("reference.R", "student_assignment.R", c("var1, var2"), c(2, 1))
}
